v3.0.0.6

+++ MPDCtrl-Desktop3/MPDCtrl/ViewModels/Classes/Profile.cs
@@ -83,8 +83,8 @@ namespace MPDCtrl.ViewModels.Classes
             }
         }
 
-        private int _volume = 50;
-        public int Volume
+        private double _volume = 50;
+        public double Volume
         {
             get { return _volume; }
             set
